+++ CHANGELOG.md
@@ -4,14 +4,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
 ## [Unreleased]
 
-## [0.0.3]
+## [0.0.4]
 ### Added
 - Introduced Prompt to unpaid users.
 - Adding ExtensionPay to collect payments.
 
 ## [Released]
+
+## [0.0.3]
+### Added
+- Removed calculate button, run calcs on plugin open.
+- Added percentages to the sliders.
+- On missing rent estimation, we remove the slider and show a rent input field
 
 ## [0.0.2] - 2021-09-17
 ### Changed